requirements.txt
PyYAML==6.0.2
jsonschema==4.23.0
requests==2.32.3
mkdocs==1.6.1
mkdocs-material==9.5.40
python-slugify==8.0.4
